refactor: GameManager and GameEngine
