+++ backend-ping-pong/src/duel/MatchManager.ts
@@ -1,11 +1,11 @@
 import { Player } from "./Player";
-import { PingPong } from "./PingPong";
 import { FastifyInstance } from "fastify";
 import { AmqpMessageBrocker } from "./AmqpMessageBrocker";
+import { GameManager } from "./GameManager";
 
 export class MatchManager {
     private waitingPlayers: Player[] = [];
-    private games: Map<string, PingPong> = new Map();
+    private games: Map<string, GameManager> = new Map();
     private requiredPlayers: number;
 
     constructor(private fastify: FastifyInstance, requirePlayers: number) {
@@ -20,7 +20,7 @@ export class MatchManager {
         if (this.waitingPlayers.length >= this.requiredPlayers) {
             const playerForMatch = this.waitingPlayers.splice(0, this.requiredPlayers);
             const messageBroker = new AmqpMessageBrocker(this.fastify);
-            const game = new PingPong(playerForMatch, messageBroker);
+            const game = new GameManager(playerForMatch, messageBroker);
             playerForMatch.forEach(player => {
                 player.game = game;
             });
