+++ backend-ping-pong/src/duel/Player.ts
@@ -1,12 +1,12 @@
+import { GameManager } from "./GameManager";
 import { KeyState } from "./KeyState";
-import { PingPong } from "./PingPong";
 import { v4 as uuidv4 } from 'uuid'
 
 export class Player {
     public id: string;
     public ws: any;
     public keyState: KeyState;
-    public game: PingPong | null = null;
+    public game: GameManager | null = null;
 
     constructor(ws: any) {
         this.id = 'player-' + uuidv4();
